pyproject.toml
[build-system]
requires = ["wheel", "setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "stracker"
version = "0.0.0"
description = "Simple Experiment Tracker (STracker)"
readme = "README.md"
authors = [{name = "Bai Huanyu", email = "eiphnix@gmail.com"}]
license = {file = "LICENSE"}

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["configuration", "config", "dictionary", "nested", "argument parsing"]
dependencies = ["pyyaml", "naapc>='1.1.0'", "numpy", "matplotlib", "aim"]
requires-python = ">=3.7"

[project.optional-dependencies]
dev = ["black", "bumpver", "twine", "isort", "pyinstrument", "tqdm"]

[project.urls]
repository = "https://github.com/eiphy/stracker"

[tool.bumpver]
current_version = "0.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"', 'version = "{version}"'
]
"src/stracker/__init__.py" = ["{version}"]
